% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pack.R
\name{bin_pack_ffd}
\alias{bin_pack_ffd}
\title{1D bin packing "First Fit (Decreasing)" algorithm}
\usage{
bin_pack_ffd(x, cap, sort = TRUE)
}
\arguments{
\item{x}{A numeric vector of values to be fit into bins.}

\item{cap}{Bin capacity in units of values in \code{x}. A scalar value.}

\item{sort}{Determines whether the input vector should be sorted in decreasing
order before applying the "First Fit" algorithm ("First Fit Decreasing").}
}
\value{
A integer vector of labels of the same length as \code{x}. The integer
label at position \code{i} determines the assignment of the \code{i}th item
with weight \code{x[i]} to a bin.
}
\description{
1D bin packing "First Fit (Decreasing)" algorithm
}
